move classes out of main file
